requirements.txt
# Core dependencies
Pillow>=9.0.0
pytesseract>=0.3.8
pdf2image>=1.16.0
PyMuPDF>=1.19.0
ocrmypdf>=13.0.0
python-docx>=0.8.11
beautifulsoup4>=4.9.3
chardet>=4.0.0
pandas>=1.3.0
reportlab>=3.6.8

# Email handling
extract-msg>=0.28.7
email-validator>=1.1.3

# Media handling
moviepy>=1.0.3
SpeechRecognition>=3.8.1
pydub>=0.25.1

# Document handling
python-pptx>=0.6.21
pywin32>=305
striprtf>=0.0.12

# Excel handling
openpyxl>=3.0.9

# Utility
tqdm>=4.62.3

cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
colorclass==2.2.2
compressed-rtf==1.0.7
cryptography==44.0.2
Deprecated==1.2.18
deprecation==2.1.0
easygui==0.98.3
ebcdic==1.1.1
img2pdf==0.6.0
lark==1.1.9
lxml==5.3.2
markdown-it-py==3.0.0
mdurl==0.1.2
msoffcrypto-tool==5.4.2
numpy==2.2.5
olefile==0.47
oletools==0.60.2
packaging==25.0
pcodedmp==1.2.6
pdfminer.six==20250416
pi_heif==0.22.0
pikepdf==9.7.0
pycparser==2.22
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
red-black-tree-mod==1.22
rich==14.0.0
RTFDE==0.1.2
six==1.17.0
soupsieve==2.7
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
win_unicode_console==0.5
wrapt==1.17.2
opencv-python==4.9.0.80
